requirements.txt
python-telegram-bot==21.0
pyyaml==6.0.1
